rename TwitchEventSub to EventSub. Note that some of the dependendencies are still using old file names
