+++ src/Core/Commands/Definitions/JoinChatCommands.cs
@@ -31,14 +31,14 @@ public class JoinChatCommands(ICoStreamChannelsRepo coStreamChannelsRepo) : ICom
     {
         if (context.Source is not TwitchChat twitchChat)
             return new CommandResult { Response = "Having the bot join your channel is not supported from this chat" };
-        TwitchEventSubChat.JoinResult result = await twitchChat.TwitchEventSubChat.Join(context.Message.User.Id);
+        EventSubChat.JoinResult result = await twitchChat.EventSubChat.Join(context.Message.User.Id);
         string response = result switch
         {
-            TwitchEventSubChat.JoinResult.Ok => "Successfully joined channel",
-            TwitchEventSubChat.JoinResult.NotEnabled => "Channel joining is currently not enabled",
-            TwitchEventSubChat.JoinResult.AlreadyJoined => "Already joined",
-            TwitchEventSubChat.JoinResult.UserNotFound => "You don't exist according to Twitch",
-            TwitchEventSubChat.JoinResult.StreamOffline => "You are not live",
+            EventSubChat.JoinResult.Ok => "Successfully joined channel",
+            EventSubChat.JoinResult.NotEnabled => "Channel joining is currently not enabled",
+            EventSubChat.JoinResult.AlreadyJoined => "Already joined",
+            EventSubChat.JoinResult.UserNotFound => "You don't exist according to Twitch",
+            EventSubChat.JoinResult.StreamOffline => "You are not live",
         };
         return new CommandResult { Response = response };
     }
@@ -47,11 +47,11 @@ public class JoinChatCommands(ICoStreamChannelsRepo coStreamChannelsRepo) : ICom
     {
         if (context.Source is not TwitchChat twitchChat)
             return new CommandResult { Response = "Having the bot leave your channel is not supported from this chat" };
-        TwitchEventSubChat.LeaveResult result = await twitchChat.TwitchEventSubChat.Leave(context.Message.User.Id);
+        EventSubChat.LeaveResult result = await twitchChat.EventSubChat.Leave(context.Message.User.Id);
         string response = result switch
         {
-            TwitchEventSubChat.LeaveResult.Ok => "Successfully left channel",
-            TwitchEventSubChat.LeaveResult.NotJoined => "Not joined",
+            EventSubChat.LeaveResult.Ok => "Successfully left channel",
+            EventSubChat.LeaveResult.NotJoined => "Not joined",
         };
         return new CommandResult { Response = response };
     }
