+++ src/Core/Core.csproj
@@ -26,7 +26,7 @@
 		<ProjectReference Include="..\Inputting\Inputting.csproj" />
 		<ProjectReference Include="..\Match\Match.csproj" />
 		<ProjectReference Include="..\PersistenceMongoDB\PersistenceMongoDB.csproj" />
-		<ProjectReference Include="..\TwitchEventSub\TwitchEventSub.csproj" />
+		<ProjectReference Include="..\EventSub\EventSub.csproj" />
 	</ItemGroup>
 
 	<ItemGroup>
