+++ src/EventSub/EventSub.csproj
