+++ src/EventSub/Notifications/ChannelChatSettingsUpdate.cs
@@ -1,6 +1,6 @@
-using TwitchEventSub.Messages;
+using EventSub.Messages;
 
-namespace TwitchEventSub.Notifications;
+namespace EventSub.Notifications;
 
 using static ChannelChatSettingsUpdate;
