+++ src/EventSub/Notifications/ChannelFollow.cs
@@ -1,7 +1,7 @@
 using NodaTime;
-using TwitchEventSub.Messages;
+using EventSub.Messages;
 
-namespace TwitchEventSub.Notifications;
+namespace EventSub.Notifications;
 
 using static ChannelFollow;
