+++ src/EventSub/Notifications/ChannelSubscribe.cs
@@ -1,7 +1,7 @@
 using System.Runtime.Serialization;
-using TwitchEventSub.Messages;
+using EventSub.Messages;
 
-namespace TwitchEventSub.Notifications;
+namespace EventSub.Notifications;
 
 using static ChannelSubscribe;
