+++ src/EventSub/Notifications/ChannelSubscriptionMessage.cs
@@ -1,7 +1,7 @@
 using System.Linq;
-using TwitchEventSub.Messages;
+using EventSub.Messages;
 
-namespace TwitchEventSub.Notifications;
+namespace EventSub.Notifications;
 
 using static ChannelSubscriptionMessage;
