+++ tests/EventSub.Tests/EventSub.Tests.csproj
@@ -1,0 +1,5 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <ItemGroup>
+        <ProjectReference Include="..\..\src\EventSub\EventSub.csproj" />
+    </ItemGroup>
+</Project>
