+++ tests/EventSub.Tests/MessageParsingTests.cs
@@ -1,9 +1,9 @@
 using NodaTime;
 using NUnit.Framework;
-using TwitchEventSub.Messages;
-using TwitchEventSub.Notifications;
+using EventSub.Messages;
+using EventSub.Notifications;
 
-namespace TwitchEventSub.Tests;
+namespace EventSub.Tests;
 
 /// <summary>
 /// Tests that the examples from the Twitch documentation page
