--- tests/TwitchEventSub.Tests/TwitchEventSub.Tests.csproj
@@ -1,5 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <ItemGroup>
-        <ProjectReference Include="..\..\src\TwitchEventSub\TwitchEventSub.csproj" />
-    </ItemGroup>
-</Project>
